add tsdoc
