+++ package.json
@@ -5,9 +5,10 @@
   "main": "dist/bundle.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "rollup -c rollup.config.js",
+    "build": "rollup -c rollup.config.js && npm run typedoc",
     "watch": "rollup -c rollup.config.js -w",
-    "prepublishOnly": "rollup -c rollup.config.js"
+    "typedoc": "typedoc --options tsdoc.json --out ./docs/ ./src/main.ts",
+    "prepublishOnly": "npm run build"
   },
   "repository": {
     "type": "git",
@@ -28,6 +29,9 @@
   "devDependencies": {
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-typescript": "^8.3.1",
-    "rollup": "^2.61.1"
+    "rollup": "~2.66",
+    "typedoc": "^0.22.13",
+    "typedoc-plugin-missing-exports": "^0.22.6",
+    "typescript": "^4.6.3"
   }
 }
