+++ rollup.config.js
@@ -1,4 +1,5 @@
-import babel from "@rollup/plugin-babel"
+import babel from "@rollup/plugin-babel";
+import typescript from "@rollup/plugin-typescript";
 import pkg from "./package.json";
 const banner = `/*!
   niconicomments.js v${pkg.version}
@@ -7,12 +8,17 @@ const banner = `/*!
 */`;
 
 export default {
-    input: 'src/main.js',
+    input: 'src/main.ts',
     output: {
         file: 'dist/bundle.js',
         format: 'umd',
         name: 'NiconiComments',
         banner
     },
-    plugins: [babel()]
+    plugins: [
+        typescript({
+            tsconfig: 'tsconfig.json',
+        }),
+        babel()
+    ]
 }
